EES-7290 split out routes to separate file
Again this removes circular dependency issues where e.g.
(AuthContext.tsx → msal.ts → routes.ts → Page Containers → Components →
AuthContext.tsx)
